requirements.txt
apache-beam==2.13.0
google-cloud-datastore==1.7.3
google-cloud-storage==1.14.0